backend/requirements.txt
# framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# database
sqlalchemy==2.0.25
alembic==1.13.1

# ml & recommendations
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
scipy==1.11.4

# llm & ai
google-generativeai==0.3.2
python-dotenv==1.0.0

# api & utilities
httpx==0.26.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# cors & security
python-cors==1.0.0

# development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14